Week-9/requirements.txt
numpy
pillow
onnxruntime
fastapi
uvicorn